refactor(timeline): split socialInteractions into reactions/comments (524→49)
socialInteractions.ts was 524 LOC — likes, dislikes, and the full comment CRUD in
one file. Extracted verbatim (behavior-preserving):
- social-shared.ts — the `db` cast + getCurrentUserId (shared internals)
- reactions.ts — toggleReaction (internal) + toggleLike, toggleDislike
- comments.ts — enrichWithProfiles (internal) + addComment, updateComment,
  deleteComment, getEventComments, getCommentReplies

socialInteractions.ts (49 LOC) keeps the cross-cutting getEventCounts and
re-exports the rest, so timeline/index.ts imports all 8 from one module — public
surface unchanged.

tsc 0 errors.
